requirements.txt
click
spotipy
python-dateutil
requests
black